Now the player cannot attack more than once. Screw you cheaters.

+++ js/game/player.js
@@ -7,51 +7,54 @@
                 let cardId = this.cardId,
                     cardObject = allCards.enemyCards.filter(obj => obj.cardId === cardId)[0];
 
-                if (!cardObject.isPlayerCard && cardObject.isPlaced && !cardObject.isJustPlaced &&
-                    localStorage.canAttack === 'true') {
-                    localStorage.canAttack = 'false';
-                    localStorage.isPlayerTurn = 'false';
-                    localStorage.hasPlayerPlacedCard = 'true';
+                if (localStorage.playerAlreadyAttacked === 'false') {
+                    if (!cardObject.isPlayerCard && cardObject.isPlaced && !cardObject.isJustPlaced &&
+                        localStorage.canAttack === 'true') {
+                        localStorage.canAttack = 'false';
+                        localStorage.isPlayerTurn = 'false';
+                        localStorage.hasPlayerPlacedCard = 'true';
 
-                    let attacker = playerCards.filter(c => c.cardId === Number(localStorage.attackerId))[0];
+                        let attacker = playerCards.filter(c => c.cardId === Number(localStorage.attackerId))[0];
 
-                    if (attacker.ability === 'stealEnemyHealth') {
-                        cardAbilities.stealFromEnemyHealth(attacker, playerAvatars);
-                        endGame.checkForEndGame(playerAvatars, allCards);
-                    } else if (attacker.ability === 'stealMana') {
-                        cardAbilities.stealManaFromEnemyPlayer(attacker, playerAvatars)
-                    } else if (attacker.ability === 'normal') {
-                        cardObject.health -= attacker.attack;
-                        cardController.performAttackAnimation(attacker, cardObject);
+                        if (attacker.ability === 'stealEnemyHealth') {
+                            cardAbilities.stealFromEnemyHealth(attacker, playerAvatars);
+                            endGame.checkForEndGame(playerAvatars, allCards);
+                        } else if (attacker.ability === 'stealMana') {
+                            cardAbilities.stealManaFromEnemyPlayer(attacker, playerAvatars)
+                        } else if (attacker.ability === 'normal') {
+                            cardObject.health -= attacker.attack;
+                            cardController.performAttackAnimation(attacker, cardObject);
 
-                        if (cardObject.health <= 0) {
-                            let indexToRemove = allCards.enemyCards.indexOf(cardObject);
+                            if (cardObject.health <= 0) {
+                                let indexToRemove = allCards.enemyCards.indexOf(cardObject);
 
-                            stage.removeChild(cardObject.cardContainer);
-                            allCards.enemyCards.splice(indexToRemove, 1);
-                            localStorage.setItem('hasToPlaceCard', 'true');
+                                stage.removeChild(cardObject.cardContainer);
+                                allCards.enemyCards.splice(indexToRemove, 1);
+                                localStorage.setItem('hasToPlaceCard', 'true');
 
-                            cardController.arrangeEnemyCardsOnField();
+                                cardController.arrangeEnemyCardsOnField();
+                            }
+
+                            if (cardObject.healthStat) {
+                                cardObject.healthStat.text = cardObject.health;
+                            }
+
+                            if (cardObject.isAvatar) {
+                                statsController.updateEnemyHealth(cardObject.health < 0 ? 0 : cardObject.health);
+                            }
+
+                            endGame.checkForEndGame(playerAvatars, allCards);
+                            localStorage.playerDamageDealt = Number(localStorage.playerDamageDealt) + Number(attacker.attack);
                         }
 
-                        if (cardObject.healthStat) {
-                            cardObject.healthStat.text = cardObject.health;
-                        }
-
-                        if (cardObject.isAvatar) {
-                            statsController.updateEnemyHealth(cardObject.health < 0 ? 0 : cardObject.health);
-                        }
-
-                        endGame.checkForEndGame(playerAvatars, allCards);
-
-                        localStorage.playerDamageDealt = Number(localStorage.playerDamageDealt) + Number(attacker.attack);
+                        localStorage.playerAlreadyAttacked = 'true';
+                    } else {
+                        cardObject.isJustPlaced = false;
                     }
-                } else {
-                    cardObject.isJustPlaced = false;
-                }
 
-                playerAvatars[1].mana = 10;
-                statsController.updateEnemyMana(10);
+                    playerAvatars[1].mana = 10;
+                    statsController.updateEnemyMana(10);
+                }
             }
 
             function initializeAttackEvent(event, playerCards) {
